+++ src/main/java/org/opensearch/knn/index/codec/derivedsource/DerivedSourceIndexOperationListener.java
@@ -19,10 +19,12 @@ import org.opensearch.core.xcontent.MediaTypeRegistry;
 import org.opensearch.core.xcontent.XContentBuilder;
 import org.opensearch.index.engine.Engine;
 import org.opensearch.index.mapper.ParseContext;
+import org.opensearch.index.mapper.SourceFieldMapper;
 import org.opensearch.index.shard.IndexingOperationListener;
 import org.opensearch.knn.index.DerivedKnnByteVectorField;
 import org.opensearch.knn.index.DerivedKnnFloatVectorField;
 import org.opensearch.knn.index.VectorDataType;
+import org.opensearch.knn.index.codec.KNN10010Codec.KNN10010DerivedSourceStoredFieldsWriter;
 import org.opensearch.knn.index.mapper.KNNVectorFieldMapperUtil;
 
 import java.io.IOException;
@@ -52,29 +54,43 @@ public class DerivedSourceIndexOperationListener implements IndexingOperationLis
             return operation;
         }
 
-        Function<Map<String, Object>, Map<String, Object>> injectTransformer = createInjectTransformer(operation);
-        if (injectTransformer == null) {
+        Pair<Function<Map<String, Object>, Map<String, Object>>> transformers = createInjectTransformer(operation);
+        if (transformers == null) {
             return operation;
         }
+        Function<Map<String, Object>, Map<String, Object>> injectTransformer = transformers.first();
+        Function<Map<String, Object>, Map<String, Object>> maskTransformer = transformers.second();
 
         Tuple<? extends MediaType, Map<String, Object>> originalSource = XContentHelper.convertToMap(
             operation.parsedDoc().source(),
             true,
             operation.parsedDoc().getMediaType()
         );
-        Map<String, Object> derivedSource = injectTransformer.apply(originalSource.v2());
-
+        Map<String, Object> cleanVectorSource = injectTransformer.apply(originalSource.v2());
         try (BytesStreamOutput bStream = new BytesStreamOutput();) {
-            XContentBuilder builder = MediaTypeRegistry.contentBuilder(originalSource.v1(), bStream).map(derivedSource);
+            XContentBuilder builder = MediaTypeRegistry.contentBuilder(originalSource.v1(), bStream).map(cleanVectorSource);
             builder.close();
             operation.parsedDoc().setSource(bStream.bytes(), XContentType.valueOf(originalSource.v1().subtype().toUpperCase(Locale.ROOT)));
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+
+        IndexableField field = operation.parsedDoc().rootDoc().getField(SourceFieldMapper.CONTENT_TYPE);
+        if (field == null || field.storedValue() == null) {
+            return operation;
+        }
+        Map<String, Object> maskedVectorSource = maskTransformer.apply(originalSource.v2());
+        try (BytesStreamOutput bStream = new BytesStreamOutput();) {
+            XContentBuilder builder = MediaTypeRegistry.contentBuilder(originalSource.v1(), bStream).map(maskedVectorSource);
+            builder.close();
+            field.storedValue().setBinaryValue(bStream.bytes().toBytesRef());
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
         return operation;
     }
 
-    private Function<Map<String, Object>, Map<String, Object>> createInjectTransformer(Engine.Index operation) {
+    private Pair<Function<Map<String, Object>, Map<String, Object>>> createInjectTransformer(Engine.Index operation) {
         Map<String, List<Object>> injectedVectors = new HashMap<>();
 
         // For each document, we get the relevant vector fields to compute the injection logic
@@ -98,11 +114,14 @@ public class DerivedSourceIndexOperationListener implements IndexingOperationLis
         }
 
         Map<String, Function<Object, Object>> injectTransformers = new HashMap<>();
+        Map<String, Function<Object, Object>> maskTransformers = new HashMap<>();
         for (Map.Entry<String, List<Object>> entry : injectedVectors.entrySet()) {
             Iterator<Object> iterator = entry.getValue().iterator();
             injectTransformers.put(entry.getKey(), (Object o) -> o == null ? o : iterator.next());
+            maskTransformers.put(entry.getKey(), (Object o) -> o == null ? o : KNN10010DerivedSourceStoredFieldsWriter.MASK);
         }
-        return XContentMapValues.transform(injectTransformers, true);
+
+        return new Pair<>(XContentMapValues.transform(injectTransformers, true), XContentMapValues.transform(maskTransformers, true));
     }
 
     private boolean isRecoverySourceEnabled(Engine.Index operation) {
@@ -115,5 +134,8 @@ public class DerivedSourceIndexOperationListener implements IndexingOperationLis
             return KNNVectorFieldMapperUtil.deserializeStoredVector(vectorBytesRef, vectorDataType);
         }
         return vectorValue;
+    }
+
+    private record Pair<T>(T first, T second) {
     }
 }